remove JsonMixin and plug the to_dict to db.Document which can filter items
